#246 optimize package card usage
