+++ modules/desktop/src/components/packages/packages.svelte
@@ -4,7 +4,7 @@
 	import type { GUIPackage } from '$libs/types';
 	import { PackageStates } from '$libs/types';
 	import Preloader from '@tea/ui/Preloader/Preloader.svelte';
-	import PackageCard from '@tea/ui/package-card/package-card.svelte';
+	import Package from "./package.svelte";
 	import { packagesStore } from '$libs/stores';
 
 	import { installPackage } from '@native';
@@ -14,15 +14,6 @@
 	export let category = ''; // filter
 
 	let packages: GUIPackage[] = [];
-
-	const getCTALabel = (state: PackageStates): string => {
-		return {
-			[PackageStates.AVAILABLE]: $t("package.install-label").toUpperCase(),
-			[PackageStates.INSTALLED]: $t("package.installed-label").toUpperCase(),
-			[PackageStates.INSTALLING]: $t("package.installing-label").toUpperCase(),
-			[PackageStates.UNINSTALLED]: $t("package.reinstall-label").toUpperCase(),
-		}[state];
-	};
 
 	packagesStore.subscribe((ps) => {
 		packages = category ?
@@ -39,11 +30,9 @@
 	{#if packages.length > 0}
 		{#each packages as pkg}
 			<div class={pkg.state === PackageStates.INSTALLING ? 'animate-pulse' : ''}>
-				<PackageCard
+				<Package
 					{pkg}
-					link={`/packages/${pkg.slug}`}
-					ctaLabel={getCTALabel(pkg.state)}
-					onClickCTA={async () => {
+					onClick={async () => {
 						try {
 							pkg.state = PackageStates.INSTALLING;
 							await installPackage(pkg.full_name);
