+++ modules/desktop/src/components/search-popup-results/search-popup-results.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 	import { searchStore } from '$libs/stores';
-	import { t } from '$libs/translations'; 
 	import type { GUIPackage } from '$libs/types';
 	import Preloader from '@tea/ui/Preloader/Preloader.svelte';
-	import PackageCard from '@tea/ui/package-card/package-card.svelte';
+	import Package from "$components/packages/package.svelte";
 	import { PackageStates } from '$libs/types';
 	import Posts from '@tea/ui/posts/posts.svelte';
 
@@ -39,15 +38,6 @@
 
 	searchStore.searching.subscribe((v) => (loading = v));
 
-	const getCTALabel = (state: PackageStates): string => {
-		return {
-			[PackageStates.AVAILABLE]: $t("package.install-label").toUpperCase(),
-			[PackageStates.INSTALLED]: $t("package.installed-label").toUpperCase(),
-			[PackageStates.INSTALLING]: $t("package.installing-label").toUpperCase(),
-			[PackageStates.UNINSTALLED]: $t("package.reinstall-label").toUpperCase(),
-		}[state];
-	};
-
 	const onClose = () => {
 		term = '';
 	};
@@ -74,11 +64,9 @@
 			{#if packages.length > 0}
 				{#each packages as pkg}
 					<div class={pkg.state === PackageStates.INSTALLING ? 'animate-pulse' : ''}>
-						<PackageCard
+						<Package
 							{pkg}
-							link={`/packages/${pkg.slug}`}
-							ctaLabel={getCTALabel(pkg.state)}
-							onClickCTA={async () => {
+							onClick={async () => {
 								try {
 									pkg.state = PackageStates.INSTALLING;
 									await installPackage(pkg.full_name);
